Add internal/image.go
The `Image` type is referenced by all ship-it components (the api,
syncd, and operator) so it needs to be defined high up in the package
heirarchy.
