+++ internal/ecrconsumer/image_values.go
@@ -2,45 +2,17 @@ package ecrconsumer
 
 import (
 	"fmt"
-	"strings"
+
+	"ship-it/internal"
 
 	"ship-it/pkg/apis/k8s.wattpad.com/v1alpha1"
 )
-
-type Image struct {
-	Registry   string
-	Repository string
-	Tag        string
-}
-
-func parseImage(repo string, tag string) (*Image, error) {
-	arr := strings.Split(repo, "/")
-
-	var registry string
-	var repository string
-	if len(arr) == 2 {
-		repository = arr[1]
-		registry = arr[0]
-	} else {
-		return nil, fmt.Errorf("malformed repo: %s", repo)
-	}
-
-	return &Image{
-		Registry:   registry,
-		Repository: repository,
-		Tag:        tag,
-	}, nil
-}
-
-func (i Image) URI() string {
-	return i.Registry + "/" + i.Repository
-}
 
 func getImagePath(obj map[string]interface{}, serviceName string) []string {
 	for key, val := range obj {
 		if key == "image" {
 			if img, ok := val.(map[string]interface{}); ok {
-				image, err := parseImage(img["repository"].(string), img["tag"].(string))
+				image, err := internal.ParseImage(img["repository"].(string), img["tag"].(string))
 				if err != nil {
 					return []string{}
 				}
@@ -75,7 +47,7 @@ func table(vals map[string]interface{}, path []string) map[string]interface{} {
 	return tabled
 }
 
-func update(vals map[string]interface{}, img Image) {
+func update(vals map[string]interface{}, img internal.Image) {
 	path := getImagePath(vals, img.Repository)
 	if len(path) == 0 {
 		return
@@ -124,7 +96,7 @@ func cleanUpMapValue(v interface{}) interface{} {
 	}
 }
 
-func WithImage(img Image, r v1alpha1.HelmRelease) v1alpha1.HelmRelease {
+func WithImage(img internal.Image, r v1alpha1.HelmRelease) v1alpha1.HelmRelease {
 	copy := r.DeepCopy()
 
 	cleanMap := cleanUpStringMap(copy.Spec.Values)
